% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChicagoCityBounds.R
\docType{data}
\name{ChicagoCityBounds}
\alias{ChicagoCityBounds}
\title{City of Chicago Boundaries}
\format{
sf with 1 observations of 5 variables:
\describe{
\item{name}{chr}
\item{objectid}{num}
\item{shape_area}{num}
\item{shape_len}{num}
\item{geometry}{sfc_MULTIPOLYGON}
}
}
\usage{
ChicagoCityBounds
}
\description{
City of Chicago Boundaries
}
\keyword{datasets}
