% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demography_1980.R
\docType{data}
\name{Demography_1980}
\alias{Demography_1980}
\title{Cook County Demographic Data - 1980}
\format{
A data frame with 1220 observations and 14 variables:
\describe{
\item{GISJOIN}{chr}
\item{NHGISST}{chr}
\item{NHGISCTY}{chr}
\item{GISJOIN2}{chr}
\item{SHAPE_AREA}{num}
\item{SHAPE_LEN}{num}
\item{TotPop80}{int Total Population of a given Tract}
\item{totpop1980Pr}{num Percent of persons living in a given tract accroding to the total population of Cook County}
\item{hisplatPr80}{num Percent of Hispanic/Latino persons living in agiven tract}
\item{whitePr80}{num Percent of White persons living in a given tract}
\item{aframPr80}{num Percent of African American/Black persons living in a given tract}
\item{asianPr80}{num Percent of Asian persons living in a given tract}
\item{AmIndAKatPr80}{num Percent of American Indian/Alaska Native persons living in a given tract}
\item{geometry}{sfc_MULTIPOLYGON}}
}
\usage{
Demography_1980
}
\description{
Cook County Demographic Data - 1980
}
\keyword{datasets}
