% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demography_2000.R
\docType{data}
\name{Demography_2000}
\alias{Demography_2000}
\title{Cook County Demographic Data - 2000}
\format{
sf with 1338 observations and 9 variables:
\describe{
\item{GEOID}{chr}
\item{TotPop}{num Total Population of a given Tract}
\item{totpop2000Pr}{num Percent of persons living in a given tract accroding to the total population of Cook County}
\item{hisplatPr00}{num Percent of Hispanic/Latino persons living in agiven tract}
\item{whitePr00}{num Percent of White persons living in a given tract}
\item{aframPr00}{num Percent of African American/Black persons living in a given tract}
\item{asianPr00}{num Percent of Asian persons living in a given tract}
\item{AmIndAKatPr00}{num Percent of American Indian/Alaska Native persons living in a given tract}
\item{geometry}{sfc_MULTIPOLYGON}}
}
\usage{
Demography_2000
}
\description{
Cook County Demographic Data - 2000
}
\keyword{datasets}
