% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demography_2010.R
\docType{data}
\name{Demography_2010}
\alias{Demography_2010}
\title{Cook County Demographic Data - 2010}
\format{
sf with 1314 observations and 9 variables:
\describe{
\item{GEOID}{chr}
\item{TotPop}{num Total Population of a given Tract}
\item{totpop2010Pr}{num Percent of persons living in a given tract accroding to the total population of Cook County}
\item{hisplatPr10}{num Percent of Hispanic/Latino persons living in agiven tract}
\item{whitePr10}{num Percent of White persons living in a given tract}
\item{aframPr10}{num Percent of African American/Black persons living in a given tract}
\item{asianPr10}{num Percent of Asian persons living in a given tract}
\item{AmIndAKatPr10}{num Percent of American Indian/Alaska Native persons living in a given tract}
\item{geometry}{sfc_MULTIPOLYGON}}
}
\usage{
Demography_2010
}
\description{
Cook County Demographic Data - 2010
}
\keyword{datasets}
