% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demography_2020.R
\docType{data}
\name{Demography_2020}
\alias{Demography_2020}
\title{Cook County Demographic Data - 2020}
\format{
sf with 1330 observations and 9 variables:
\describe{
\item{GEOID}{chr}
\item{TotPop}{num Total Population of a given Tract}
\item{totpop2020Pr}{num Percent of persons living in a given tract accroding to the total population of Cook County}
\item{hisplatPr20}{num Percent of Hispanic/Latino persons living in agiven tract}
\item{whitePr20}{num Percent of White persons living in a given tract}
\item{aframPr20}{num Percent of African American/Black persons living in a given tract}
\item{asianPr20}{num Percent of Asian persons living in a given tract}
\item{AmIndAKatPr20}{num Percent of American Indian/Alaska Native persons living in a given tract}
\item{geometry}{sfc_MULTIPOLYGON}}
}
\usage{
Demography_2020
}
\description{
Cook County Demographic Data - 2020
}
\keyword{datasets}
