% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISCOLLOQUIUM-package.R
\docType{package}
\name{GISCOLLOQUIUM-package}
\alias{GISCOLLOQUIUM}
\alias{GISCOLLOQUIUM-package}
\title{GISCOLLOQUIUM: Deindustrialization in the City of Chicago (1980-2020)}
\description{
This package contains basic demographic and income data for every decennial year between 1980 and 2020. Though it can be used for many analyses, the original intention is utilize the embeded data to look at and analyze spatial trends over the forty year period as it relates to the age of deindustrialization in the City of Chicago.
}
\author{
\strong{Maintainer}: Noel ngarcia7@uchicago.edu Garcia (\href{https://orcid.org/0000-0002-4623-0878}{ORCID})

}
\keyword{internal}
