% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Median_Household_Income_2010.R
\docType{data}
\name{Median_Household_Income_2010}
\alias{Median_Household_Income_2010}
\title{Median Household Income - 2010}
\format{
sf with 1314 rows and 6 variables:
\describe{
\item{GEOID}{chr}
\item{NAME}{chr}
\item{variable}{chr}
\item{M.HouseholdIncome10}{num Median Household Income}
\item{moe}{num margin of error for Median Household Income estimate}
\item{geometry}{sfc_MULTIPOLYGON}
}
}
\usage{
Median_Household_Income_2010
}
\description{
Median Household Income - 2010
}
\keyword{datasets}
