% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_Capita_Income_1980.R
\docType{data}
\name{Per_Capita_Income_1980}
\alias{Per_Capita_Income_1980}
\title{Per Capita Income - 1980}
\format{
sf with 1229 rows and 9 variables:
\describe{
\item{GISJOIN}{chr}
\item{NHGISST}{chr}
\item{NHGISCTY}{chr}
\item{GISJOIN2}{chr}
\item{SHAPE_AREA}{num}
\item{SHAPE_LEN}{num}
\item{YEAR}{chr}
\item{PCI}{chr Per Capita Income of the decennial year}
\item{geometry}{sfc_MULTIPOLYGON}
}
}
\usage{
Per_Capita_Income_1980
}
\description{
Per Capita Income - 1980
}
\keyword{datasets}
