% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_Capita_Income_1990.R
\docType{data}
\name{Per_Capita_Income_1990}
\alias{Per_Capita_Income_1990}
\title{Per Capita Income - 1989}
\format{
sf with 1352 rows and 10 variables:
\describe{
\item{GISJOIN}{chr}
\item{NHGISST}{chr}
\item{NHGISCTY}{chr}
\item{GISJOIN2}{chr}
\item{SHAPE_AREA}{num}
\item{SHAPE_LEN}{num}
\item{STATE}{chr}
\item{COUNTY}{chr}
\item{PCI}{chr Per Capita Income of the approximate decennial year}
\item{geometry}{sfc_MULTIPOLYGON}
}
}
\usage{
Per_Capita_Income_1990
}
\description{
Per Capita Income - 1989
}
\keyword{datasets}
