% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_Capita_Income_1999.R
\docType{data}
\name{Per_Capita_Income_1999}
\alias{Per_Capita_Income_1999}
\title{Per Capita Income - 1999}
\format{
sf with 1343 rows and 11 variables:
\describe{
\item{GISJOIN}{chr}
\item{NHGISST}{chr}
\item{NHGISCTY}{chr}
\item{GISJOIN2}{chr}
\item{SHAPE_AREA}{num}
\item{SHAPE_LEN}{num}
\item{YEAR}{int}
\item{STATE}{chr}
\item{COUNTY}{chr}
\item{PCI99}{chr Per Capita Income of the approximate decennial year}
\item{geometry}{sfc_MULTIPOLYGON}
}
}
\usage{
Per_Capita_Income_1999
}
\description{
Per Capita Income - 1999
}
\keyword{datasets}
