% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZoI.R
\docType{data}
\name{ZoI}
\alias{ZoI}
\title{ZoI - Zones of Interest}
\format{
sf with 227 observations of 28 variables:
\describe{
\item{commarea}{chr}
\item{commarea_n}{num}
\item{countyfp10}{chr}
\item{geoid10}{chr}
\item{namelsad10}{chr}
\item{notes}{chr}
\item{statefp10}{chr}
\item{tractce10}{chr}
\item{Intersecti}{num}
\item{Intersec_1}{chr}
\item{Intersec_2}{num}
\item{Intersec_3}{chr}
\item{Intersec_4}{num}
\item{Intersec_5}{num}
\item{Intersec_6}{chr}
\item{Intersec_7}{num}
\item{Intersec_8}{num}
\item{Intersec_9}{num}
\item{Intersec10}{chr}
\item{Intersec11}{num}
\item{Intersec12}{chr}
\item{Intersec13}{chr}
\item{Intersec14}{chr}
\item{Intersec15}{chr}
\item{Intersec16}{chr}
\item{Intersec17}{chr}
\item{geometry}{sfc_POLYGON}
}
}
\usage{
ZoI
}
\description{
ZoI - Zones of Interest
}
\keyword{datasets}
